[parser]: Extract DataType and Codec types into dedicated files
The column.go file had grown to contain multiple unrelated type families.
This refactoring improves code organization by separating concerns into
focused files with dedicated unit tests.

- Move CodecClause and CodecSpec to codec.go with codec_test.go
- Move DataType and 12 related types to datatype.go with datatype_test.go
- Reduce column.go to only Column, ColumnAttribute, DefaultClause, TTLClause
- Add unit tests using direct struct construction (no parser dependency)
