+++ pkg/utils/ptr.go
@@ -1,0 +1,5 @@
+package utils
+
+func Ptr[T any](v T) *T {
+	return &v
+}
